requirements.txt
absl-py==0.9.0
alembic==1.4.2
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
azure-core==1.8.0
azure-storage-blob==12.4.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
cryptography==3.0
cycler==0.10.0
databricks-cli==0.11.0
dill==0.3.2
docker==4.3.0
entrypoints==0.3
Flask==1.1.2
future==0.18.2
gast==0.3.3
gin-config==0.3.0
gitdb==4.0.5
GitPython==3.1.7
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gorilla==0.3.0
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
image-classifiers==1.0.0
importlib-metadata==1.7.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mlflow==1.10.0
msrest==0.6.18
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.0
Pillow==7.2.0
plotly==4.9.0
prometheus-client==0.8.0
prometheus-flask-exporter==0.15.4
promise==2.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-datasets==2.1.0
tensorflow-estimator==2.3.0
tensorflow-metadata==0.23.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
